Mixed topology type writing.
When writing a mixed topology, "number of elements" and "nodes per
element" could not be enough information to know the dataset size,
hyperslab size and hyperslab offset of the connectivities.
When working with Mixed topologies we have to know explicitly this
information.

Now there are some more variables in the uniform and spatial grid
descriptors to store this information (ConnectivitySize,
GlobalConnectivitySize, ConnectivitySizePerGrid, ...). As a
difference with the other required distribute data, this data is
distributed when calling to "WriteTopology" and not in the
initialization procedure.